Data.Tree.MerklePatricia

Description

An implementation of a Merkle-Patricia tree, with additions and deletions of elements from root hashes.

Note that this module is not meant to be used as a handful data-structures for classic operations; Those operations are better done with a Map or a HashMap containers. A MerklePatriciaTree can however be leveraged to produce proofs which sizes are logarithmic in the size of the tree, and which can be verified using only a root hash. This makes is feasible to embed verifiable operations on an existing tree with only tiny proofs.

Types

data MerklePatriciaTree alg a

An opaque MerklePatriciaTree, parameterized by a hash algorithm and a type for its values.

data Proof alg

An opaque Proof which can be used to prove membership or atomic operations such as add or delete

Constructing

fromList :: forall alg a. (HashAlgorithm alg, Serialise a) => [(ByteString, a)] -> MerklePatriciaTree alg a

Construct a Merkle-Patricia Tree from a list of (key, value) pairs.

This operation is expensive for it computes intermediate hashes of each nodes and leaves of the tree.

properties:

>>> ∀xs. toList (fromList xs) == xs

toList :: forall alg a. MerklePatriciaTree alg a -> [(ByteString, a)]

Convert a Merkle-Patricia Tree into a list of (key, value) pairs.

properties:

>>> ∀xs. toList (fromList xs) == xs

Analyzing

root :: forall alg a. MerklePatriciaTree alg a -> Digest alg

Access the root hash of the MerklePatriciaTree. Two trees with the same root hash are equal.

null :: MerklePatriciaTree alg a -> Bool

True if the given MerklePatriciaTree is empty.

properties:

>>> ∀mpt. (size mpt == 0) == (null mpt)
True

size :: MerklePatriciaTree alg a -> Int

Count the number of values stored in the tree (that is, number of non-empty leaves). Note that this does not necessarily correspond to the number of nodes in the tree.

We have however the following property:

properties:

>>> ∀xs. size (fromList xs) == length xs
True

depth :: MerklePatriciaTree alg a -> Int

Count the maximum number of levels in the tree.

Proof

Constructing

mkProof

Arguments

:: ByteString

-> MerklePatriciaTree alg a
A key pointing to one element in the tree

-> Maybe (Proof alg)
The target tree

Construct a proof for a given key of the MerklePatriciaTree. Returns Nothing if the key does not exists. The proof is compact and does not information embedded in the element itself. This means that, the corresponding (key, value) is necessary for any use of the proof.

Using

member

Arguments

:: forall alg a. (Serialise a, HashAlgorithm alg)

=> (ByteString, a)
The (key, value) to check

-> Digest alg
The root of the MerklePatriciaTree with (supposedly) the (key, value)

-> Proof alg
A proof for the (key, value) in the corresponding MerklePatriciaTree

-> Bool

Check whether an element exists in the MerklePatriciaTree.

properties:

>>> ∀mpt, ∀(k, v) ∈ mpt. member (k, v) (root mpt) <$> mkProof k mpt
Just True

>>> ∀mpt, ∀(k, v) ∈ mpt, ∃k' / k ≠ k'. member (k', v) (root mpt) <$> mkProof k mpt
Just False

>>> ∀mpt, ∀(k, v) ∈ mpt, ∃v' / v ≠ v'. member (k, v') (root mpt) <$> mkProof k mpt
Just False

add

Arguments

:: forall alg a. (Serialise a, HashAlgorithm alg)

=> (ByteString, a)
The (key, value) to add.

-> Digest alg
The root of the MerklePatriciaTree without the (key, value)

-> Proof alg
A proof for the (key, value) in the resulting MerklePatriciaTree

-> Maybe (Digest alg)
Returns, the root of the MerklePatriciaTree with the (key, value) added, or Nothing if the proof is invalid for the given root.

Check whether an element exists in the MerklePatriciaTree, yielding the root hash of the new tree.

properties:

>>> ∀mpt, ∀(k, v) ∈ mpt, add (k, v) (root (mpt \ (k,v))) <$> mkProof k mpt
Just (root mpt)

>>> ∀mpt, ∀(k, v) ∈ mpt, ∃v' / v ≠ v'. add (k, v') (root (mpt \ (k,v))) <$> mkProof k mpt
Nothing

delete

Arguments

:: forall alg a. (Serialise a, HashAlgorithm alg)

=> (ByteString, a)
The (key, value) to delete.

-> Digest alg
The root of the MerklePatriciaTree with the (key, value)

-> Proof alg
A proof for the (key, value) in the original MerklePatriciaTree

-> Maybe (Digest alg)
Returns, the root of the MerklePatriciaTree without the (key, value), or Nothing if the (key, value) is not present in the original tree.

Delete an element from a MerklePatriciaTree, yielding the root hash of the new tree.

properties:

>>> ∀mpt, ∀(k, v) ∈ mpt, delete (k, v) (root mpt) <$> mkProof k mpt
Just (root (mpt \ (k, v)))

>>> ∀mpt, ∀(k, v) ∈ mpt, ∃k' / k ≠ k'. delete (k', v) (root mpt) <$> mkProof k mpt
Nothing

>>> ∀mpt, ∀(k, v) ∈ mpt, ∃v' / v ≠ v'. delete (k, v') (root mpt) <$> mkProof k mpt
Nothing

>>> ∀mpt, ∀(k, v) ∈ mpt, let Just p = mkProof k mpt. add (k, v) <$> delete (k, v) (root mpt) p <*> pure p
Just (root mpt)

Debugging

proofSize :: forall alg. Proof alg -> Int

Size of a Proof. Returns the number of hashes needed for the proof.

pretty :: forall alg a. Show a => MerklePatriciaTree alg a -> String

Pretty-print a MerklePatriciaTree in ascii, such that levels are aligned horizontally (one identation-level per level in the tree), and prefixes are positioned such that they are visually easy to follow. The outputs look like this:

*--- Root hash
|                      *---- Values at the leaves
|                      |
*-> #3801f:            |
    #ee41c: \0         |
    #16944:  \0      vvvvv
    #9d288:   \001100 ↦ ()
    #d0bd1:   \100110 ↦ ()
    #38e90:  \1000110 ↦ ()
    #978d8: \1
    #5ac65:  \0
    #6f2de:   \00
    #4fe20:     \0110 ↦ ()
    #d1489:     \1100 ↦ ()
    #d0bd1:   \100110 ↦ ()
    #22882:  \100
    #4fe20:   ^ \0110 ↦ ()
    #d1489:   | \1100 ↦ ()
     ^^^^^    |   ^^^
       |      |    |
       |      |    *--- Leaf or node's prefix
       |      |
       |      *---- Branch bit
       |
       |
       |
       *----- Nodes / Leaves' hashes

Useful for debugging / reasonning.

Re-Exports

Serialise

class Serialise a

Types that are instances of the Serialise class allow values to be quickly encoded or decoded directly to a CBOR representation, for object transmission or storage.

Since: serialise-0.2.0.0

serialise :: Serialise a => a -> ByteString

Serialise a Haskell value to an external binary representation.

The output is represented as a lazy ByteString and is constructed incrementally.

Since: serialise-0.2.0.0

deserialise :: Serialise a => ByteString -> a

Deserialise a Haskell value from the external binary representation (which must have been made using serialise or related function).

Throws: DeserialiseFailure if the given external representation is invalid or does not correspond to a value of the expected type.

Since: serialise-0.2.0.0

HashAlgorithm

class HashAlgorithm a

Class representing hashing algorithms.

The interface presented here is update in place and lowlevel. the Hash module takes care of hidding the mutable interface properly.

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

Blake2b_160

data Blake2b_160

Blake2b (160 bits) cryptographic hash algorithm

Blake2b_224

data Blake2b_224

Blake2b (224 bits) cryptographic hash algorithm

Blake2b_256

data Blake2b_256

Blake2b (256 bits) cryptographic hash algorithm